pyproject.toml
[project]
name = "tiling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "cppimport>=22.8.2",
    "evn>=0.7.8",
    "ipython>=9.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pybind11>=3.0.1",
    "scikit-video>=1.1.11",
    "torch>=2.9.0",
]
